+++ packages/adminui-vue/components/controls/UDropdown/UDropdown.vue
@@ -12,10 +12,6 @@
       <slot />
     </div>
 
-    <transition
-      name="dropdown-transition"
-      @enter="beforeEnter"
-    >
       <div
         v-show="visible && $slots.dropdown"
         :key="renderKey"
@@ -35,7 +31,6 @@
           <slot name="dropdown" />
         </div>
       </div>
-    </transition>
   </div>
 </template>
 
@@ -147,12 +142,14 @@ export default {
 
   beforeDestroy () {
     this.$refs.dropdown.remove()
+    this.disconnectObserver()
   },
 
   methods: {
     toggleVisible () {
       if (this.disabled) return
       this.visible = !this.visible
+      this.beforeEnter(this.$refs.dropdown)
     },
 
     beforeEnter (el) {
@@ -188,6 +185,7 @@ export default {
         }
       )
       this.checkAndUpdatePopupPosition(this.popperInstance)
+      if (this.observer) return
       // set watcher for observe changes width and height popup when user change content in him
       const callback = () => this.checkAndUpdatePopupPosition()
       this.observer = new MutationObserver(callback)
@@ -219,7 +217,7 @@ export default {
     close () {
       this.visible = false
       this.$emit('close')
-      if (this.observer) this.observer.disconnect()
+      this.disconnectObserver()
     },
 
     closeByEscape (event) {
@@ -250,6 +248,10 @@ export default {
         bottom: y,
         left: x
       })
+    },
+    disconnectObserver () {
+      if (this.observer) this.observer.disconnect()
+      this.observer = null
     }
   }
 }